# lib/api-zod/src/generated/types/gameChargeEntryBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mother Bot API - SKZ Financial Hub
 * OpenAPI spec version: 0.2.0
 */

export interface GameChargeEntryBody {
  telegramId: string;
  gameId: string;
  /** Amount in SKZ to charge for game entry */
  amount: string;
}