# lib/api-zod/src/generated/types/gameCreditRewardBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mother Bot API - SKZ Financial Hub
 * OpenAPI spec version: 0.2.0
 */

export interface GameCreditRewardBody {
  telegramId: string;
  /** ID of the charge transaction returned by /internal/game/charge-entry */
  chargeTransactionId: number;
  /** Signed token from /internal/game/validate-result proving game was played */
  resultToken: string;
  /** Player's score for this game session */
  score?: number;
}