# lib/api-zod/src/generated/types/gameCreditRewardResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mother Bot API - SKZ Financial Hub
 * OpenAPI spec version: 0.2.0
 */

export interface GameCreditRewardResponse {
  success: boolean;
  transactionId: number;
  newSkzBalance: string;
  commissionDeducted: string;
  netRewarded: string;
}