# lib/api-zod/src/generated/types/gameValidateResultBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mother Bot API - SKZ Financial Hub
 * OpenAPI spec version: 0.2.0
 */

export interface GameValidateResultBody {
  telegramId: string;
  /** ID of the charge transaction returned by /internal/game/charge-entry */
  chargeTransactionId: number;
  /** Player's final score — server validates against stored minWinScore and minDurationMs */
  score: number;
}