# lib/api-zod/src/generated/types/gameValidateResultResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mother Bot API - SKZ Financial Hub
 * OpenAPI spec version: 0.2.0
 */

export interface GameValidateResultResponse {
  ok: boolean;
  /** Signed HMAC token (10-min TTL) required by /internal/game/credit-reward */
  resultToken: string;
}