# lib/api-zod/src/generated/types/internalWithdrawBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mother Bot API - SKZ Financial Hub
 * OpenAPI spec version: 0.2.0
 */
import type { InternalWithdrawBodyDestination } from './internalWithdrawBodyDestination';

export interface InternalWithdrawBody {
  telegramId: string;
  /** Withdrawal method (e.g. "ton") */
  methodCode: string;
  /** Amount of SKZ to withdraw */
  amountSkz: string;
  /** Method-specific destination details (e.g. { tonAddress }) */
  destination?: InternalWithdrawBodyDestination;
}