# lib/api-zod/src/generated/types/starsInvoiceResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mother Bot API - SKZ Financial Hub
 * OpenAPI spec version: 0.2.0
 */

export interface StarsInvoiceResponse {
  ok: boolean;
  invoiceLink: string;
  payload: string;
  expectedSkz: string;
}