# lib/api-zod/src/generated/types/tonDepositIntentBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mother Bot API - SKZ Financial Hub
 * OpenAPI spec version: 0.2.0
 */

export interface TonDepositIntentBody {
  telegramId: string;
  /** Amount of TON to deposit */
  amountTon: number;
}