# lib/api-zod/src/generated/types/tonDepositIntentResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mother Bot API - SKZ Financial Hub
 * OpenAPI spec version: 0.2.0
 */

export interface TonDepositIntentResponse {
  ok: boolean;
  intentId: number;
  memo: string;
  depositAddress: string;
  amountTon: number;
  expectedSkz: string;
  expiresAt: Date;
}